Rework InteractionBlock
The implementation from this commit is suboptimal in many ways, but it
is simple and allowed me to update the unit tests with few minor
changes.
The current implementation will most likely be replaced in the near
future, but no changes to the test cases should be required (in theory).