Greatly improve performance of iteration at high resolutions (<1000bp)
Our previous implementation performed quite poorly at high resolutions,
as we were processing one row at a time while paying the overhead of
fetching the block index and data every row.
This was done with the intention of minimizing the amount of read-ahead
we do, as well as getting pixels in the correct order without doing any
explicit sort.

However this was too slow.
The current solution is less clean but performance is much better (15x
at 10bp on some of our datasets).
Instead of processing one row at a time, we now process rows in chunks.
Chunk sizes are computed as a fraction of chromosome sizes, and thus
grow linearly with resolution, making the overhead to process a chunk
comparable across resolution.

Another benefit of this approach is that indexing of InteractionBlock is
no longer needed: sorting pixels is enough.