Properly enforce block cache size in hic files
Cache size is now always expressed in bytes.
The only place where size is still expressed as number of pixels, is in
code responsible of evicting blocks to keep the cache size below
capacity.

Furthermore, the lower bound for cache size is now enforced right before
setting the capacity, not when computing the space required by each
submatrix.