Bugfix
It turns out we cannot simply divide bin_ids to generate coarse
matrices. The reason being that when the number of bins in a matrix is
not a multiple of the coarsening factor, the last and first pixel(s) of
adjacent chromosomes could be mapped to the same coarse bin (which of
course doesn't make sense).