Support reading hic v6 and v7
hic v7 support was tested using the following file:
https://www.encodeproject.org/files/ENCFF718AWL/

hic v6 support was not tested as I am unable to find a file this old.

I was also unable to acquire a copy of juicer_tools (or
juicebox_tools/hic_tools) old enough to be able to generate hic v7 or
v6.