Address segfault when merging iterators of hic pixels
The segfault occurred when storing a relatively large number of
iterators in a vector.
If the vector was created by repeatedly calling emplace_back(), and one
or more iterators were generated immediately after, the pointer to the
hic::PixelSelector stored in hic::PixelSelector::iterator
would become dangling.

The segfault is solved by storing member variables of hic::PixelSelector
as shared pointers, then passing these pointers when creating an
instance of hic::PixelSelector::iterator<>.