Implement HiCBlockPartitioner class
This class splits a stream of pixels into chunks, keeping track of
which chunks map to which blocks in the matrix-matrix grid.
After calling finalize() on an HiCBlockPartitioner object, the class can
be used to merge chunks mapping to the same matrix-matrix block.

The functionality provided by this class is meant to be used to
preprocess a stream of pixels to aid the incremental creation of .hic files.