Avoid needlessly recomputing norm. expected values
This is in theory an easy feature to support.
However it in practice .hic files can have multiple entries for the same
normalization vector/normalized expected values.
I assume such files could be generated by calling juicer_tools addNorm
multiple times.
This PR assumes that the last vector for a given normalization method and unit
in a file is the most recent (and the one that users actually care about).