Properly support copying weights from .mcool files with small bin tables
Fix a bug that could occur when converting .mcool with small genomes and
coarse resolutions to .hic.
When the resolution is sufficiently low, the bin table for small genomes
will consist of one bin per chromosome.
When creating .cool files, cooler flags the bin table for such files as
tables with variable bin size.
This is achieved by setting bin_size to NULL in the HDF5 file.
hictk represents resolutions with integers, and cannot thus represent
null values. Instead, when opening a .cool file with a table with
variable bins, hictk returns 0 as file resolution.
.hic files with BP unit do not support tables of variable bin sizes, thus
we have to treat bin tables described above as fixed bin tables (i.e.
tables with a resolution other than 0).

This commit implements a workaround to for the above issue.