Support linking to Arrow::python instead of Arrow::arrow_shared
This is mostly useful for hickpy, where linking to Arrow::arrow_shared
causes linking to a specific version of Arrow (which depends on
the specific version of Pyarrow that was installed when building
hictkpy wheels).
This is not desirable, because it forces us to pin the version of pyarrow
required by hictkpy (which could easily result in unresolvable environments).
Here we assume that libarrow_python.so links to some version of
libarrow.so, which when loaded brings in all symbols required by
hictk and hictkpy.
It should be noted that Arrow has very few  guarantees when it comes to ABI
stability (see https://github.com/apache/arrow/issues/41707).
This means that hictkpy needs to test all possible combinations of pyarrow/arrow
versions on all supported platforms and architectures to make sure there
are no ABI changes (at least to the parts that hictk and hictkpy depend on).