Optimize performance of inter-chromosomal queries on Cooler files.
This PR introduces two changes aimed at improving the performance of trans queries on Cooler files.
Both optimizations improve the performance of calls to cooler::Dataset::iterator<T>::jump_to_col()
by reducing the number of chunks that need to be fetched and processed.

The first optimization involves keeping track of the number of values at the beginning of rows
that have been skipped before finding the first pixel overlapping with the given query.
This number is updated every time jump_to_col() is called.

The second optimization is an optimization around binary search (std::lower_bound to be more precise).
The idea is that when calling std::lower_bound from within calling jump_to_col(), we know that bin2_ids
are very often (approximately) uniformly distributed.
Thus, we can tweak lower_bound() to be smarter when picking the pivot point (instead of always picking
the mid-point between first and last).
This makes a big difference in our case because not all lookups performed by lower_bound take the same time:
lookups that require loading a new chunk are way more expensive than performing lookups of values
that have already been prefetched.
Thus, I re-implemented std::lower_bound to take this into consideration.
Basically, given an iterator it, we don't directly check if *it >= value.
Instead, we first check if the given value is between chunk[0] and chunk[-1], where chunk
is the buffer underlying it, and chunk[0] and chunk[-1] are the first and last values overlapping
the given query.

So, for every call to lower_bound (including recursive calls), we do the following:
- check if *first >= value (return first if true)
- check *(last - 1) <= value (return last if true)
- check if value could be in the chunk underlying first (call std::lower_bound on the chunk if true)
- check if value could be in the chunk underlying last (call std::lower_bound on the chunk if true)
- pick a pivot
- check if value could be in the chunk underlying pivot (call std::lower_bound on the chunk if true)
- if value <= pivot_chunk[0], keep looking to the left, else keep looking to the right

Basically, we take advantage of the fact that once we call *it we can access any value from its chunk
pretty much for free.