Workaround weird bug in exception handling
For some reason some exceptions were not caught (as in they
reached main and caused std::abort() to be called) because of the
if constexpr inside std::visit.
The odd part is that this only occurred when building the Dockerfile
(I was not able to reproduce locally, nor with the regular CI).